Module modularise photofs
